+++ src/main/java/com/appland/appmap/process/hooks/ToggleRecord.java
@@ -8,17 +8,14 @@ import com.appland.appmap.process.ExitEarly;
 import com.appland.appmap.record.ActiveSessionException;
 import com.appland.appmap.record.IRecordingSession;
 import com.appland.appmap.record.Recorder;
+import com.appland.appmap.reflect.HttpServletRequest;
+import com.appland.appmap.reflect.HttpServletResponse;
+import com.appland.appmap.reflect.FilterChain;
 import com.appland.appmap.transform.annotations.*;
+import com.appland.appmap.util.Logger;
 
 import java.io.IOException;
 import java.io.PrintWriter;
-
-import javax.servlet.FilterChain;
-import javax.servlet.ServletException;
-import javax.servlet.ServletRequest;
-import javax.servlet.ServletResponse;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
 
 /**
  * Hooks to toggle event recording. This could be either via HTTP or by entering a unit test method.
@@ -39,7 +36,7 @@ public class ToggleRecord {
     } catch (ActiveSessionException e) {
       res.setStatus(HttpServletResponse.SC_NOT_FOUND);
     } catch (IOException e) {
-      System.err.printf("failed to write response: %s\n", e.getMessage());
+      Logger.printf("failed to write response: %s\n", e.getMessage());
     }
   }
 
@@ -55,7 +52,7 @@ public class ToggleRecord {
       writer.write(responseJson);
       writer.flush();
     } catch (IOException e) {
-      System.err.printf("failed to write response: %s\n", e.getMessage());
+      Logger.printf("failed to write response: %s\n", e.getMessage());
     }
   }
 
@@ -69,14 +66,17 @@ public class ToggleRecord {
     }
   }
 
-  @ExcludeReceiver
-  @HookClass("javax.servlet.http.HttpServlet")
-  public static void service(Event event, HttpServletRequest req, HttpServletResponse res)
-      throws ExitEarly {
+  private static void service(Object[] args) throws ExitEarly {
+    if (args.length != 2) {
+      return;
+    }
+
+    final HttpServletRequest req = new HttpServletRequest(args[0]);
     if (!req.getRequestURI().endsWith(RecordRoute)) {
       return;
     }
 
+    final HttpServletResponse res = new HttpServletResponse(args[1]);
     if (req.getMethod().equals("GET")) {
       doGet(req, res);
     } else if (req.getMethod().equals("POST")) {
@@ -88,25 +88,50 @@ public class ToggleRecord {
     throw new ExitEarly();
   }
 
-  @ContinueHooking
-  @ExcludeReceiver
-  @HookClass("javax.servlet.Filter")
-  public static void doFilter(Event event,
-                              ServletRequest req,
-                              ServletResponse res,
-                              FilterChain chain)
-                              throws IOException, ServletException, ExitEarly {
-    if (!(req instanceof HttpServletRequest)) {
+  private static void skipFilterChain(Object[] args) throws ExitEarly {
+    if (args.length != 3) {
       return;
     }
 
-    if (!((HttpServletRequest) req).getRequestURI().endsWith(RecordRoute)) {
+    final HttpServletRequest req = new HttpServletRequest(args[0]);
+    if (!req.getRequestURI().endsWith(RecordRoute)) {
       return;
     }
 
-    chain.doFilter(req, res);
+    final FilterChain chain = new FilterChain(args[2]);
+    chain.doFilter(args[0], args[1]);
 
     throw new ExitEarly();
+  }
+
+  @ExcludeReceiver
+  @ArgumentArray
+  @HookClass("javax.servlet.http.HttpServlet")
+  public static void service(Event event, Object[] args) throws ExitEarly {
+    service(args);
+  }
+
+  @ExcludeReceiver
+  @ArgumentArray
+  @HookClass(value = "jakarta.servlet.http.HttpServlet", method = "service")
+  public static void serviceJakarta(Event event, Object[] args) throws ExitEarly {
+    service(args);
+  }
+
+  @ContinueHooking
+  @ExcludeReceiver
+  @ArgumentArray
+  @HookClass("javax.servlet.Filter")
+  public static void doFilter(Event event, Object[] args) throws ExitEarly {
+    skipFilterChain(args);
+  }
+
+  @ContinueHooking
+  @ExcludeReceiver
+  @ArgumentArray
+  @HookClass(value = "jakarta.servlet.Filter", method = "doFilter")
+  public static void doFilterJakarta(Event event, Object[] args) throws ExitEarly {
+    skipFilterChain(args);
   }
 
   private static void startTest(Event event) {
@@ -140,7 +165,7 @@ public class ToggleRecord {
 
       recorder.start(fileName, metadata);
     } catch (ActiveSessionException e) {
-      System.err.printf("AppMap: %s\n", e.getMessage());
+      Logger.printf("AppMap: %s\n", e.getMessage());
     }
   }
 
@@ -148,7 +173,7 @@ public class ToggleRecord {
     try {
       recorder.stop();
     } catch (ActiveSessionException e) {
-      System.err.printf("AppMap: %s\n", e.getMessage());
+      Logger.printf("AppMap: %s\n", e.getMessage());
     }
   }