+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.0] - 2020-06-29
+### Added
+- Support for Jakarta
+
+### Removed
+- `javax.servlet` dependency
+
+### Changed
+- All logging is disabled by default. To enable logging use the `appmap.debug` system property.
+- Loading times have been reduced significantly.
+
 ## [0.1.0] - 2020-06-05
 ### Added
 - Name property added to appmap metadata